requirements.txt
# MAX_AI Scanner Service - Dependencies

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client
httpx>=0.25.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Audio (optional - falls back to system if not available)
# playsound>=1.3.0
# pygame>=2.5.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Type checking
mypy>=1.7.0

# Formatting
black>=23.11.0
isort>=5.12.0
